verbs: Introduce counters read verb
The user supplies counters instance and a reference to an output
array of uint64_t.
The driver reads the hardware counters values and writes them to
the output index location in the user supplied array.
All counters values are represented as uint64_t types.

The read counter values in the array are defined according to the
configuration defined by the user in the ibv_attach_counters_point_xxx
functions.

Signed-off-by: Raed Salem <raeds@mellanox.com>
Signed-off-by: Yishai Hadas <yishaih@mellanox.com>