mlx5: Allocate huge page chunks only when needed
When huge pages allocation is enabled and user requests to create a
raw packet QP with sq of size 0 (only rq) we first allocate a new huge
page for the rq buffer.

When allocating the sq buffer the driver will first check if there are
already allocated huge pages that have available space within them and
try to allocate chunks from that page instead of getting a new one. In
this case it will find the page we allocated for the rq but since the
requested size is 0, it will not consume or mark any chunks of memory
within that page but the address and reference to that page is still
stored in that buffers context.

During the release of the buffers, we first release the rq buffer and
since all used chunks of memory are cleared upon the rq buffer release,
we also free the page and remove it from the available pages list.
When the driver attempts to release the sq buffer, it will try to
fetch the page information from the pages list which will cause access
to a null pointer since the page was already removed from the list.

Therefore, this fix deals with cases of allocation within huge page
where requested size is 0 by not performing the flow above.

Fixes: 8c4791ae2395 ("libmlx5: First version of libmlx5")
Fixes: ec90e3c2f8b9 ("Allocate separate RQ and SQ buffers for Raw Packet QP")
Signed-off-by: Ariel Levkovich <lariel@mellanox.com>
Signed-off-by: Yishai Hadas <yishaih@mellanox.com>