verbs: Add IDRs array attribute type to ioctl() interface
Methods sometimes need to get a flexible set of idrs and not a strict
set as can be achieved today by the conventional idr attribute. This
is an idrs-array-like behavior.
Since this may be popular used, we add a new IDRS_ARRAY attribute to
the generic uverbs ioctl layer.
This attribute is embedded in methods, like any other attributes we
currently have.

We align the user-space code with the respected changes in the kernel.

Signed-off-by: Guy Levi <guyle@mellanox.com>
Signed-off-by: Yishai Hadas <yishaih@mellanox.com>