verbs: Do not add device whose uverbs device is not accessible
When RDMA devices are used in container environment, orchestration
software may choose to not map a uverbs character device to container.

Adding an RDMA device which is not accessible for any verbs operation,
but only accessible via sysfs is not very useful to real world
applications but it just creates confusion.

Therefore, do not add RDMA devices to ib device list, if its respective
character device is not accessible.

Signed-off-by: Parav Pandit <parav@mellanox.com>
Signed-off-by: Yishai Hadas <yishaih@mellanox.com>