verbs: Added new rates derived from 2x and HDR support
Due to 2x and HDR addition new rates exposed:
28 Gbps - FDR 2x.
50 Gbps - EDR 2x or HDR 1x.
400 Gbps - HDR 8x.
600 Gbps - HDR 12x.

Utility methods were updated to support the new rates.
Rate to mult - Convert the IB rate enum to a multiple of the base rate
of 2.5 Gbit/sec.  For example, IBV_RATE_5_GBPS will be converted to 2,
since 5 Gbit/sec is 2 * 2.5 Gbit/sec.

Rate to mbps - Convert IB rate enum to the mbps value. conversion done
by multiplication of IB rate in mbps from IB spec and width needed to reach
the rate. For example, 400 GBPS is reached by HDR 8x so the mbps rate is
53.125 Gbit/sec * 8 = 425,000 mbps.

In addition, speed_str() and width_str() as part of ibv_devinfo were
updated to consider the new HDR and 2x options.

Signed-off-by: Michael Guralnik <michaelgur@mellanox.com>
Signed-off-by: Yishai Hadas <yishaih@mellanox.com>