Add Microsoft Azure RDMA device to MLX4 HCA table
Microsoft Azure supports Mellanox MLX4 RDMA devices in HPC VM SKUs. The device
is virtualized and exposed over VMBUS, and registered as a MLX4 Infininband
hardware in RDMA layer.

This devcie is on VMBUS, not on PCI tree. Add it to the HCA table so it can
be properly discovered by user-mode driver.

Signed-off-by: Long Li <longli@microsoft.com>