librdmacm: Allow RDMA CM use with connected QP when it is not managing the QP
The RDMA CM documentation indicates it supports being used for communication
management only (I.E. application manages QP). This change enables using the
RDMA CM for connected QP communication management without it managing the QP.
Since the Connection Response event is consumed in the library on the active
side, the application must ensure it closes the race of transitioning the QP
to RTR.

This change only impacts the current error path when QP are not managed
and is consistent with the other unmanaged QP checks done throughout the code.

Signed-off-by: Steve Welch <swelch@systemfabricworks.com>
Signed-off-by: James Swaro <jswaro@cray.com>