verbs: If the uverbs module is not loaded allow get_devices to retry init
Currently if uverbs is not loaded libibverbs latches into a permanent
disabled state on the first try to ibv_get_device_list. This is not
compatible with hot-plug which could insert the uverbs module at any
later point. Recheck for uverbs presence until we get a successful init
or a hard failure during init.

Cc: stable@linux-rdma.org
Fixes: c632072027a8 ("verbs: Refresh cahced ibv_device list")
Signed-off-by: Jason Gunthorpe <jgg@mellanox.com>