Generate complete and sanitized static libraries
Once rdma-core started to introduce ccan and util internal libraries the
static libraries stopped working in a sane way. We need to include the
internal libraries within the static ones and sanitize the symbol names to
get things back to where they used to be.

Add a post-processing script that figures out what to do and invokes
binutils to generate clean static libraries.

Signed-off-by: Jason Gunthorpe <jgg@mellanox.com>