verbs: Add RDMA_STATIC_PROVIDERS
This macro is to be set by the user to a comma separated list of providers
that should be included in the static link. The special word 'all' will
include every supported provider.

To make this work we list all of the provider libraries in the pkg-config
file, if the macro isn't used then none of them will contribute to the
link, as before.

Signed-off-by: Jason Gunthorpe <jgg@mellanox.com>