mlx5: Add packet based credit mode support
The device can support two credit modes, message based (default) and
packet based. In order to enable packet based the QP should be created
with special flag that indicates this.

This patch adds support for the new DV QP creation flag that can be used
for RC QPs in order to change the credit mode from message based to
packet based.

Signed-off-by: Danit Goldberg <danitg@mellanox.com>
Signed-off-by: Yishai Hadas <yishaih@mellanox.com>